move feed loading into `PostTracker`
