+++ Mlem/Models/Trackers/Post Tracker.swift
@@ -10,11 +10,36 @@ import Foundation
 @MainActor
 class PostTracker: ObservableObject {
     
-    @Published var page: Int = 1
     @Published var isLoading: Bool = true
     @Published private(set) var posts: [APIPostView] = .init()
     
+    private var page: Int = 1
     private var ids: Set<Int> = .init()
+    
+    /// A method to request the tracker loads the next page of posts
+    /// - Parameters:
+    ///   - account: The `SavedAccount` for the logged in user
+    ///   - communityId: An optional `Int` if you are retrieving posts for a specific community
+    ///   - sort: The sorting type for the feed
+    ///   - type: The type of feed the tracker should load
+    func loadNextPage(account: SavedAccount, communityId: Int?, sort: SortingOptions?, type: FeedType) async throws {
+        let request = GetPostsRequest(
+            account: account,
+            communityId: communityId,
+            page: page,
+            sort: sort,
+            type: type
+        )
+        
+        let response = try await APIClient().perform(request: request)
+        
+        guard !response.posts.isEmpty else {
+            return
+        }
+        
+        add(response.posts)
+        page += 1
+    }
     
     /// A method to add new posts into the tracker, duplicate posts will be rejected
     /// - Parameter newPosts: The array of `APIPostView` you wish to add
